implement momentum strategy

+++ data/stock.py
@@ -1,27 +1,22 @@
 from jqdatasdk import *
-import pandas as pd
-import datetime
 import os
+from util import *
 
 auth_flag = False
 
 
-def auth_check(func):
-    def _wrapper(*args, **kwargs):
-        global auth_flag
-        if not auth_flag:
-            auth(os.getenv("JQUSERNAME"), os.getenv("JQPASSWORD"))
-        res = func(*args, **kwargs)
-        return res
-    return _wrapper
+def auth_check():
+    global auth_flag
+    if not auth_flag:
+        auth(os.getenv("JQUSERNAME"), os.getenv("JQPASSWORD"))
 
 
-@auth_check
 def get_stock_list():
     """
     获取所有A股股票列表，上交所.XSHG，深交所.XSHE
     :return: stock_list
     """
+    auth_check()
     stock_list = list(get_all_securities(["stock"]).index)
     return stock_list
 
@@ -34,18 +29,27 @@ def get_local_stock_list():
     return stock_list
 
 
-@auth_check
 def get_index_list(index_symbol='000300.XSHG'):
     """
     获取指数成分股，指数代码查询：https://www.joinquant.com/indexData
     :param index_symbol: 指数的代码，默认沪深300
     :return: list，成分股代码
     """
-    stocks = get_index_stocks(index_symbol)
+    file_path = os.path.join(os.path.dirname(__file__), "index_list", index_symbol)
+    if os.path.exists(file_path):
+        stocks = json.load(open(file_path))
+        print(f"从本地读取指数 {index_symbol} 股票列表")
+        print(f"指数 {index_symbol} 中含有 {len(stocks)} 支股票")
+    else:
+        auth_check()
+        stocks = get_index_stocks(index_symbol)
+        with open(file_path, "w") as f:
+            f.write(json.dumps(stocks))
+        print(f"缓存指数 {index_symbol} 股票列表到本地")
+        print(f"指数 {index_symbol} 中含有 {len(stocks)} 支股票")
     return stocks
 
 
-@auth_check
 def get_single_price(code, frequency, start_date=None, end_date=None, export=False):
     """
     获取单个股票行情数据
@@ -53,12 +57,15 @@ def get_single_price(code, frequency, start_date=None, end_date=None, export=Fal
     :param frequency: 数据周期
     :param start_date: 起始日期
     :param end_date: 结束日期
+    :param export: 是否储存数据到本地
     :return: data
     """
     if start_date is None:
+        auth_check()
         start_date = get_security_info(code).start_date
     if end_date is None:
         end_date = datetime.datetime.today()
+    auth_check()
     data = get_price(code, start_date=start_date, end_date=end_date, frequency=frequency, panel=False)
     data.index.names = ["date"]
     if export:
@@ -77,17 +84,18 @@ def get_data_path(code, type):
     return file_path
 
 
-def export_data(data, code, type):
+def export_data(data, code, type, overwrite=False):
     """
     导出股票相关数据
     :param data: 股票数据
     :param code: 股票代码
     :param type: 股票数据类型：price, finance
+    :param overwrite: 是否重写所有数据
     :return: None
     """
     file_path = os.path.join(os.path.dirname(__file__), type, f"{code}.csv")
     data.index.names = ["date"]
-    if os.path.exists(file_path):
+    if os.path.exists(file_path) and not overwrite:
         existing_data = import_data(code, type)
         concat = pd.concat([existing_data, data]).drop_duplicates().sort_index()
         concat.to_csv(file_path)
@@ -103,6 +111,7 @@ def import_data(code, type, start_date=None, end_date=None):
     :param type: 股票数据类型：price, finance
     :param start_date: 起始日期
     :param end_date: 结束日期
+    :param use_cols: 截取列名
     :return: None
     """
     file_path = os.path.join(os.path.dirname(__file__), type, f"{code}.csv")
@@ -114,25 +123,56 @@ def import_data(code, type, start_date=None, end_date=None):
         data = data.loc[data.index >= start_date]
     if end_date is not None:
         data = data.loc[data.index <= end_date]
+    alert = True
+    if end_date is None:
+        end_date = datetime.datetime.today()
+    latest_possible_trading_date = get_latest_possible_trading_day(end_date)
+    if timestamp_to_string(data.index[-1]) >= latest_possible_trading_date:
+        alert = False
+    if alert:
+        print(f"{code} {type} data might be obsolete!")
     return data
 
 
 def transfer_price_frequency(data, frequency):
     """
     转换股票数据周期：开盘价（周期第一天），收盘价（周期最后一天），最高价（周期内最高价），最低价（周期内最低价）
+    如当天所有数据均为空值 NaN，则被舍弃（长假期等非交易日导致所有股票均无交易）
     :param data: 股票数据
     :param frequency: 数据周期
     :return: trans
     """
+    default_cols = ["open", "close", "high", "low", "volume", "money"]
+    multi_index = False
     trans = pd.DataFrame()
-    trans["open"] = data["open"].resample(frequency).first()
-    trans["close"] = data["close"].resample(frequency).last()
-    trans["high"] = data["high"].resample(frequency).max()
-    trans["low"] = data["low"].resample(frequency).min()
+    for col in data.columns:
+        col_name = col
+        if not isinstance(col, str) and not isinstance(col, tuple):
+            raise Exception("Unhandled DataFrame column structure!")
+        if isinstance(col, tuple):
+            multi_index = True
+            for c in col:
+                if c in default_cols:
+                    col_name = c
+                    break
+        if col_name == "open":
+            trans[col] = data[col].resample(frequency).first()
+        elif col_name == "close":
+            trans[col] = data[col].resample(frequency).last()
+        if col_name == "high":
+            trans[col] = data[col].resample(frequency).max()
+        if col_name == "low":
+            trans[col] = data[col].resample(frequency).min()
+        if col_name == "volume":
+            trans[col] = data[col].resample(frequency).sum()
+        if col_name == "money":
+            trans[col] = data[col].resample(frequency).sum()
+    if multi_index:
+        trans.columns = pd.MultiIndex.from_tuples(list(trans.columns))
+    trans = trans.dropna(how="all")
     return trans
 
 
-@auth_check
 def get_single_finance(code, date=None, statDate=None):
     """
     获取单个股票财务指标
@@ -141,11 +181,11 @@ def get_single_finance(code, date=None, statDate=None):
     :param statDate: 财报统计的季度或年份
     :return: data
     """
+    auth_check()
     data = get_fundamentals(query(indicator).filter(indicator.code == code), date=date, statDate=statDate)
     return data
 
 
-@auth_check
 def get_single_valuation(code, date=None, statDate=None):
     """
     获取单个股票估值指标
@@ -154,6 +194,7 @@ def get_single_valuation(code, date=None, statDate=None):
     :param statDate: 财报统计的季度或年份
     :return: data
     """
+    auth_check()
     data = get_fundamentals(query(valuation).filter(valuation.code == code), date=date, statDate=statDate)
     return data
 
@@ -168,23 +209,43 @@ def calculate_change_percent(data):
     return data
 
 
-@auth_check
-def update_daily_price(code, type='price', local=True):
+def update_daily_price(code, type="price", local=True, fetch=False):
     """
     更新单个股票行情数据
     :param code: 股票代码
     :param type: 股票数据类型：price, finance
+    :param local: 是否只更新数据库中存在的股票数据
+    :param fetch: 重新获取股票数据
     :return: None
     """
+    if fetch:
+        data = get_single_price(code, "daily")
+        export_data(data, code, "price", overwrite=True)
+        print(f"股票数据已重新获取：{code}")
+        return
     file_path = get_data_path(code, type)
     if os.path.exists(file_path):
-        start_date = pd.read_csv(file_path, usecols=["date"])["date"].iloc[-1]
-        today = datetime.datetime.today().strftime("%Y-%m-%d")
-        if today > start_date:
-            data = get_single_price(code, "daily", start_date, today)
-            export_data(data, code, "price")
+        update = False
+        overwrite = False
+        latest_possible_trading_day = get_latest_possible_trading_day(datetime.datetime.today())
+        try:
+            end_date = pd.read_csv(file_path, usecols=["date"])["date"].iloc[-1]
+            if end_date < latest_possible_trading_day:
+                update = True
+        except IndexError:
+            update = True
+            end_date = None
+            overwrite = True
+        if update:
+            data = get_single_price(code, "daily", end_date, latest_possible_trading_day)
+            export_data(data, code, "price", overwrite=overwrite)
             print(f"股票数据已更新： {code}")
     elif not local:
         data = get_single_price(code, "daily")
         export_data(data, code, "price")
-        print(f"股票数据已更新：{code}")
+        print(f"股票数据已添加：{code}")
+
+
+if __name__ == "__main__":
+    index_list = get_index_list()
+    print(len(index_list))
